+++ packages/adminui-vue/adminui-vue-types.js
@@ -35,7 +35,7 @@
  * @property {boolean} [sortable] Allow ordering records by values of the column.
  * @property {boolean} [filterable]
  *   If set false, the column cannot be filtered. Otherwise, default and optional custom filters will be applied
- * @property {boolean} [isLookup=false] If is set true will check attribute associatedEntity
+ * @property {boolean} [isLookup=false] If true - check attribute associatedEntity
  *   and loads description attribute for displayed value
  * @property {'SUM'|'MIN'|'MAX'|'AVG'|null} [summaryAggregationOperator]
  *   Specify aggregation operator to use, when calculate a summary for a column.