+++ packages/adminui-vue/components/UTableEntity/filter-templates/mixinForFilter.js
@@ -4,25 +4,25 @@ module.exports = {
       type: [String, Number, Boolean, Array]
     }
   },
-  created() {
-    if (this.defaultValue === undefined) return;
-    this.value = this.defaultValue;
+  created () {
+    if (this.defaultValue === undefined) return
+    this.value = this.defaultValue
     if (Array.isArray(this.defaultValue)) {
-      this.valueFrom = this.defaultValue[0];
-      this.valueTo = this.defaultValue[1];
+      this.valueFrom = this.defaultValue[0]
+      this.valueTo = this.defaultValue[1]
     }
   },
   watch: {
     isEmpty: {
       immediate: true,
-      handler(newValue) {
-        this.$emit("search-disabled", newValue);
+      handler (newValue) {
+        this.$emit('search-disabled', newValue)
       }
     }
   },
   methods: {
-    submitHandler() {
-      this.$emit("search", this.getCondition());
+    submitHandler () {
+      this.$emit('search', this.getCondition())
     }
   }
-};
+}