+++ packages/ub-pub/CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
  - nodeJS compatibility: transport.js, account that `readystatechange` event might be raised after "done" state
    The `xmlhttprequest` module, which used in NodeJS environment, in case of errors, such as socket closed,
    might fire `readystatechange` with state "DONE" (4) twice.
+ - Allow to pass `{timeout: 0}` to xhr - prevent overriding 0 value with default timeout
 
 ## [5.9.22] - 2021-12-07
 ### Fixed