+++ packages/ub-pub/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Added
+ - `UBConnection` will handle a `403 Forbidden` server-side error response and raise a UB.UBAbort message
+ instead of `InternalServerError`. This response occurs when a user tries to retrieve data from the server but fails
+ the server's data access validation with status code 403.
 
 ### Changed