+++ packages/ub-pub/CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 
 ### Changed
+ - `UTableEntity` - emitSelectedEvent will include only rows from current page (`curSelected`) instead of all 
+  rows used ever select (`selectionCache`)
 
 ### Deprecated