+++ packages/ub-pub/CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Removed
 
 ### Fixed
+
+## [5.22.2] - 2022-02-08
+### Fixed
  - restore `change password` functionality (brok by 403 code interceptor added in 5.22.0)
 
 ## [5.22.1] - 2022-01-14